// src/types/charger.ts
export type Charger = {
  addr: string;
  bnm: string;
  busiCall: string;
  busiId: string;
  busiNm: string;
  chgerId: string;
  chgerType: string;
  delDetail: string;
  delYn: string;
  kind: string;
  kindDetail: string;
  lastTedt: string;
  lastTsdt: string;
  lat: string;
  limitDetail: string;
  limitYn: string;
  lng: string;
  location: string;
  method: string;
  note: string;
  nowTsdt: string;
  output: string;
  parkingFree: string;
  powerType: string;
  stat: string;
  statId: string;
  statNm: string;
  statUpdDt: string;
  trafficYn: string;
  useTime: string;
  zcode: string;
  zscode: string;
};